% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_RC_rules.R
\name{append_RC_rules}
\alias{append_RC_rules}
\title{Append Recording Confidence Rules}
\usage{
append_RC_rules(classifications_df, longitude, latitude)
}
\arguments{
\item{classifications_df}{A dataframe containing classifications. This dataset assumes the data is in the format obtained from the classification pipeline on JASMIN.}

\item{longitude}{Longitude of the deployment.}

\item{latitude}{Latitude of the deployment.}
}
\value{
The updated classifications dataframe with appended RC rules.
}
\description{
This function appends columns onto a classifications dataframe indicating whether the classification is sensible regarding
spatial presence (1 or 0), phenology (1 or 0), and appends a column with a scale regarding identification difficulty (from 1 - 4, distinct to requiring dissection, or 5, meaning Ungraded).
The function uses preloaded 'rules' datasets provided by the NBN Record Cleaner. The function iterates through all detected top-n classifications
(e.g., \code{top_1_species}, \code{top_2_species}, etc.) and appends spatial presence, phenology, and identification difficulty for each classification separately.
}
\details{
The data appended by this function is obtained from the NBN Record Cleaner.
If you use this function or the datasets in publications, please cite the NBN Record Cleaner.
}
\references{
National Biodiversity Network (NBN) Record Cleaner.
See \url{https://nbn.org.uk/tools-and-resources/nbn-record-cleaner/}.
}
