% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_device_operation_period.R
\name{calculate_device_operation}
\alias{calculate_device_operation}
\title{Calculate number of operational nights (/days)}
\usage{
calculate_device_operation(dataframe)
}
\arguments{
\item{dataframe}{Results dataframe}
}
\value{
A number for how many nights the system collected data during the deployment period
}
\description{
Calculate number of operational nights (/days)
}
