% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_df.R
\docType{data}
\name{classification_df}
\alias{classification_df}
\title{A subset of data from a survey with the Automated Monitoring of Insects System in the UK}
\usage{
classification_df
}
\description{
This is a subset of the whole dataset, using the first 100 rows.
}
