% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications_df.R
\docType{data}
\name{classifications_df}
\alias{classifications_df}
\title{A subset of data from a survey with the Automated Monitoring of Insects System in the UK}
\usage{
classifications_df
}
\description{
This is a subset of the whole dataset, using the first 100 rows.
}
