% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_csv_files.R
\name{concat_csvs}
\alias{concat_csvs}
\title{Concatenate csv files}
\usage{
concat_csvs(dir)
}
\arguments{
\item{dir}{A string specifying the name of the directory containing the csvs to concatenate}
}
\value{
Location of where concatenated csv has been saved
}
\description{
This function concatenates all csv files in a given directory into 1 csv file
For example, the moth classifier pipeline returns one results CSV per day. We
want to combine all these together for analysis purposes.
}
