% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_credentials.R
\name{create_credentials}
\alias{create_credentials}
\title{Generate AWS Credentials JSON File}
\usage{
create_credentials(
  access_key_id,
  secret_access_key,
  file_path = "credentials.json"
)
}
\arguments{
\item{access_key_id}{A string containing the AWS Access Key ID.}

\item{secret_access_key}{A string containing the AWS Secret Access Key.}

\item{file_path}{A string specifying the file path where the JSON file should be saved. Default is "credentials.json".}
}
\value{
The function writes a JSON file to the specified location and returns the file path.
}
\description{
This function creates a JSON file containing AWS credentials and endpoint details.
}
\examples{
create_credentials("my_access_key", "my_secret_key")

# Custom file path
generate_aws_json("my_access_key", "my_secret_key", "custom_path.json")
}
