% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_check.r
\name{credentials_check}
\alias{credentials_check}
\title{Validate the existence of a credentials file and warn about security risks}
\usage{
credentials_check(credentials_path)
}
\arguments{
\item{credentials_path}{Character. The path to the credentials file.}
}
\value{
NULL (invisibly). The function stops execution if the credentials file does not exist.
}
\description{
This function checks whether the specified credentials file exists. If it does not,
an error is raised. If it does, a warning is issued about the potential risks of sharing
credentials or using them outside of the EntoInsights package.
}
