% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_snapshots.R
\name{crop_and_save_images}
\alias{crop_and_save_images}
\title{Crops snapshot images}
\usage{
crop_and_save_images(
  downloaded_files,
  classifications_df,
  output_dir = "./crops"
)
}
\arguments{
\item{bucket}{A string with the bucket name, which is the iso3 code, in lowercase (e.g., "gbr").}

\item{deployment_id}{A vector of strings with a length equal to the number of files, and recording the deployment id of each image.}

\item{data_type}{A string with the type of data (one from: "snapshot_images", "audible_recordings", "ultrasound_recordings")}

\item{filename}{A vector of strings with a length equal to the number of files, and recording the file base name of each image.}

\item{classification_id}{This is an ID that you assign which allows you to safely map the species classifcation to the downloaded files given some classifications may share the same file.}
}
\value{
A dataframe that allows you to track successful file downloads. The dataframe contains columns deployment_id, containing the deployment id, the local installation path, and the classification id.
}
\description{
This function can be used to crop object store files from the object store service. This function is useful for users intending to download subsets of data from a deployment or from multiple deployments.
To download greater numbers of files from the object store (e.g., all files in a deployment), please see the function here ().
Please note, to run this function you need to generate a credentials JSON file, which requires a key, managed by the object store data manager.
Once you have a key, you can generate a credentials object using the function, create_credentials().
}
