% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object_store_files.R
\name{download_object_store_files}
\alias{download_object_store_files}
\title{Download files from an S3 bucket}
\usage{
download_object_store_files(
  bucket,
  deployment_id,
  data_type,
  filename,
  classification_id = NULL,
  download_path = "./downloads"
)
}
\arguments{
\item{bucket}{A character string specifying the name of the S3 bucket.}

\item{deployment_id}{A character vector of deployment IDs.}

\item{data_type}{A character string specifying the type of data (used to construct the S3 key).}

\item{filename}{A character vector of filenames corresponding to the deployment IDs.}

\item{classification_id}{A character vector of classification IDs, must have the same length as \code{deployment_id} and \code{filename}.}

\item{download_path}{A character string specifying the local directory where files should be downloaded. Default is "./downloads".}
}
\value{
A dataframe containing \code{deployment_id}, \code{local_path}, and \code{classification_id} for the downloaded files.
}
\description{
This function downloads files from an AWS S3 bucket based on deployment ID,
filename, and classification ID. It ensures the directory structure exists,
supports parallel downloading, and returns a dataframe with the local file paths.
}
\details{
\itemize{
\item The function reads AWS credentials from a JSON file (\code{./credentials.json}).
\item It uses the \code{paws} package to interact with AWS S3.
\item Parallel downloading is enabled using the \code{furrr} package.
\item If any downloads fail, an error message is displayed, and the function proceeds with the remaining downloads.
}
}
