% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object_store_files.R
\name{download_top_activity_audio}
\alias{download_top_activity_audio}
\title{Download top N audio recordings with most activity}
\usage{
download_top_activity_audio(
  dataframe,
  n = 3,
  data_type = "audible_recordings",
  download_path = "./bird_downloads",
  credentials_path = "./credentials.json",
  save_download_log = TRUE
)
}
\arguments{
\item{dataframe}{Data frame containing inference results.}

\item{n}{Integer. Number of recordings to return. Default = 3.}

\item{data_type}{Character. Data type folder in bucket
e.g., audible_recordings.}

\item{download_path}{Character. Local download directory.}

\item{credentials_path}{Character. Path to AWS credentials JSON.}

\item{save_download_log}{Logical. Whether to save download log.}
}
\value{
Character vector of local file paths.
}
\description{
Identifies recordings with the highest number of detections
(based on number of rows per file) and downloads the audio files
from the object store.
}
\details{
This function  calls /code{/link{download_object_store_files}} internally.
}
\seealso{
download_object_store_files
}
