% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object_store_files.R
\name{download_top_detection_images}
\alias{download_top_detection_images}
\title{Download top N images with most detections}
\usage{
download_top_detection_images(
  dataframe,
  n = 3,
  data_type,
  download_path = "./downloads",
  credentials_path = "./credentials.json",
  save_download_log = TRUE
)
}
\arguments{
\item{dataframe}{Data frame containing inference results.}

\item{n}{Integer. Number of top images to download. Default = 3.}

\item{data_type}{Character. Data type folder in bucket
e.g., snapshot_images, motion_images, ultrasound_recordings, or audible_recordings.}

\item{download_path}{Character. Local download directory.}

\item{credentials_path}{Character. Path to AWS credentials JSON.}

\item{save_download_log}{Logical. Whether to save download log.}
}
\value{
Data frame of download results.
}
\description{
Identifies the images with the highest number of unique detections
(based on \code{crop_status}) and downloads the original image files
from the object store.
}
\details{
This function  calls \code{\link{download_object_store_files}} internally.
}
\seealso{
download_object_store_files
}
