% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntoInsights-package.R
\docType{package}
\name{EntoInsights-package}
\alias{EntoInsights}
\alias{EntoInsights-package}
\title{EntoInsights}
\description{
Provides a range of functions to support users of the Automated Monitoring of Insects (AMI) and LepiSense systems. Its functionality broadly assists users who wish to analyse classification datasets, as well as partners reviewing and retrieving their uploaded image and acoustic datasets.
}
\author{
\strong{Maintainer}: Grace Skinner \email{GraSki@ceh.ac.uk}

Other contributors:
\itemize{
  \item Dylan Carbone \email{dylcar@ceh.ac.uk} [contributor]
}

}
