% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_confidence_threshold.R
\name{filter_by_conf_threshold}
\alias{filter_by_conf_threshold}
\title{Filter the data according to a specified minimum species classifier confidence threshold.}
\usage{
filter_by_conf_threshold(dataframe, conf_threshold = 0.7)
}
\arguments{
\item{dataframe}{Results dataframe}

\item{conf_threshold}{Value between 0 and 1 to set as the confidence threshold to filter the results. Defaults to 0.7.}
}
\value{
A filtered dataframe with only predictions over a certain confidence
}
\description{
Only rows with species predictions will be included. Images with no detections or rows with detections of images other than moths will
be filtered out at this stage.
}
