% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_summary_stats.R
\name{get_activity_summary}
\alias{get_activity_summary}
\title{Generate summary stats (mean, min, max) for the number of detections and species per day/night over the deployment.}
\usage{
get_activity_summary(dataframe, detections = TRUE, species = TRUE)
}
\arguments{
\item{dataframe}{Results dataframe containing classifications.}

\item{detections}{Logical. If TRUE (default), return summary stats on detections (mean, min, max number of detections per day/night).}

\item{species}{Logical. If TRUE (default), return summary stats on species (mean, min, max number of species per day/night).}
}
\value{
List containing the specified summary statistics
}
\description{
Generate summary stats (mean, min, max) for the number of detections and species per day/night over the deployment.
}
