% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list.R
\name{get_species_list}
\alias{get_species_list}
\title{Generate a unique species list for species detected over the deployment.}
\usage{
get_species_list(dataframe, conf_threshold = 0.7)
}
\arguments{
\item{dataframe}{Results dataframe}

\item{conf_threshold}{Value between 0 and 1 to set as the confidence threshold to filter the results. Defaults to 0.7.}
}
\value{
A species list
}
\description{
Only include species predicted with a confidence over a specified threshold.
}
