% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_species_functions.R
\name{get_top_species_nights}
\alias{get_top_species_nights}
\title{Get top species ranked by nights detected}
\usage{
get_top_species_nights(dataframe, n = 10)
}
\arguments{
\item{dataframe}{Results dataframe.}

\item{n}{Number of species to return (default 10).}
}
\value{
Dataframe of top species ranked by nights detected.
}
\description{
Get top species ranked by nights detected
}
\details{
This is a helper function that calls \code{\link{summarise_species_activity}} internally.
Use this function to find species that were consistently detected across multiple nights, regardless of total counts.
}
\seealso{
\code{\link{summarise_species_activity}}, \code{\link{get_top_species_total}}
}
