% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_species_functions.R
\name{get_top_species_total}
\alias{get_top_species_total}
\title{Get top species ranked by total detections}
\usage{
get_top_species_total(dataframe, n = 10)
}
\arguments{
\item{dataframe}{Results dataframe.}

\item{n}{Number of species to return (default 10).}
}
\value{
Dataframe of top species ranked by total detections.
}
\description{
Get top species ranked by total detections
}
\details{
This is a helper function that calls \code{\link{summarise_species_activity}} internally.
Use this function when you want to quickly get the species most frequently detected over the deployment period.
}
\seealso{
\code{\link{summarise_species_activity}}, \code{\link{get_top_species_nights}}
}
