% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling_functions.R
\name{keep_only_moths}
\alias{keep_only_moths}
\title{Keep only moths from a dataset}
\usage{
keep_only_moths(data, label_column = "binary_label")
}
\arguments{
\item{data}{A data frame containing the dataset to filter.}

\item{label_column}{The name of the column containing the labels. Defaults to "binary_label".}
}
\value{
A data frame containing only rows where the binary_label is "moth".
}
\description{
This function filters rows in a dataset to keep only those with a binary label of "moth".
}
\examples{
keep_only_moths(data)
}
