% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Create a Leaflet map showing technology locations (adaptable so can input more than one set of lat/lons for multiple technologies)}
\usage{
map(latitude, longitude, labels = NULL)
}
\arguments{
\item{latitude}{Numeric vector. Latitude(s) of the technology/ies.}

\item{longitude}{Numeric vector. Longitude(s) of the technology/ies.}

\item{labels}{Optional character vector. Labels to show in popups for each marker. Defaults to "AMI system".}
}
\value{
A leaflet map object
}
\description{
Create a Leaflet map showing technology locations (adaptable so can input more than one set of lat/lons for multiple technologies)
}
