% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moths_vs_non_moths.R
\name{moths_only}
\alias{moths_only}
\title{Filter the data  to only include predictions where either the moth/non-moth OR order classifier predicts its a moth with relatively high confidence (>0.8)}
\usage{
moths_only(dataframe, conf_threshold = 0.8)
}
\arguments{
\item{dataframe}{Results dataframe}

\item{conf_threshold}{Numeric value between 0 and 1 giving the confidence threshold used to filter the data (default 0.8).}
}
\value{
A filtered dataframe containing predictions where classifiers are confident it is a moth
}
\description{
Filter the data  to only include predictions where either the moth/non-moth OR order classifier predicts its a moth with relatively high confidence (>0.8)
}
