% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity_graphs.R
\name{plot_activity_graph}
\alias{plot_activity_graph}
\title{Generate graph illustrating activity (detections or species) over the deployment period}
\usage{
plot_activity_graph(dataframe, type)
}
\arguments{
\item{dataframe}{Results dataframe containing classifications.}

\item{type}{Character string specifying what to plot.
\code{"detections"} will show the number of detected individuals per night,
\code{"species"} will show the number of unique species detected per night.}
}
\value{
ggplot graph of activity (detections or species) over the deployment period
}
\description{
Generate graph illustrating activity (detections or species) over the deployment period
}
