% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_device_operation_period.R
\name{plot_device_operation_barplot}
\alias{plot_device_operation_barplot}
\title{Create a figure showing technology operational period and how many files per night}
\usage{
plot_device_operation_barplot(dataframe)
}
\arguments{
\item{dataframe}{Results dataframe}
}
\value{
A ggplot barplot of nightly file counts e.g., number of images
}
\description{
Create a figure showing technology operational period and how many files per night
}
