% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_device_operation_period.R
\name{plot_device_operation_calendar}
\alias{plot_device_operation_calendar}
\title{Create a 3rd option for a figure showing technology operational period using calendar plot}
\usage{
plot_device_operation_calendar(dataframe)
}
\arguments{
\item{dataframe}{Results dataframe}
}
\value{
An openair calendar plot of device operation
}
\description{
Create a 3rd option for a figure showing technology operational period using calendar plot
}
