% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_device_operation_period.R
\name{plot_device_operation_tileplot}
\alias{plot_device_operation_tileplot}
\title{Create a 2nd option for a figure showing technology operational period (simpler but less detailed)}
\usage{
plot_device_operation_tileplot(dataframe)
}
\arguments{
\item{dataframe}{Results dataframe}
}
\value{
A ggplot tileplot of nightly file counts e.g., number of images
}
\description{
Create a 2nd option for a figure showing technology operational period (simpler but less detailed)
}
