% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moths_vs_non_moths.R
\name{plot_moths_vs_nonmoths}
\alias{plot_moths_vs_nonmoths}
\title{Plot a bar graph of the proportion of insects classified as moths vs non-moths. Choose whether to use
the moth/non-moth classifier and/or the insect taxonomic order classifier.}
\usage{
plot_moths_vs_nonmoths(
  dataframe,
  moth_nonmoth_classifier = TRUE,
  order_classifier = TRUE,
  conf_threshold = 0.8
)
}
\arguments{
\item{dataframe}{Results dataframe}

\item{moth_nonmoth_classifier}{TRUE or FALSE depending on whether users wants plot for the moth vs non-moth classifier}

\item{order_classifier}{TRUE or FALSE depending on whether users wants plot for the order classifier}

\item{conf_threshold}{Numeric value between 0 and 1 giving the confidence threshold used to split moth classifications
into high- and low-confidence groups (default 0.8).}
}
\value{
A  barplot showing the proportion of moths to non-moths in the dataset
}
\description{
Plot a bar graph of the proportion of insects classified as moths vs non-moths. Choose whether to use
the moth/non-moth classifier and/or the insect taxonomic order classifier.
}
