% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_species_functions.R
\name{plot_top_species_bubble}
\alias{plot_top_species_bubble}
\title{Plot activity of top species over time using bubble-style plot}
\usage{
plot_top_species_bubble(dataframe, type = c("detections", "nights"), n = 3)
}
\arguments{
\item{dataframe}{Results dataframe}

\item{type}{Character. Determines which species are plotted: "detections" or "nights" (default "detections")
\itemize{
\item "detections": plot species with the highest total number of detections over the deployment.
\item "nights": plot species detected on the most nights (regardless of how many detections per night).
}}

\item{n}{Number of species to include (default 3)}
}
\value{
ggplot object showing species activity over time
}
\description{
Plot activity of top species over time using bubble-style plot
}
\details{
Bubble size always represents the number of detections per night.
This function internally calls \code{\link{summarise_species_activity}} to select the top species for plotting.
}
\seealso{
\code{\link{summarise_species_activity}}, \code{\link{get_top_species_total}}, \code{\link{get_top_species_nights}}
}
