% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_species_functions.R
\name{plot_top_species_line}
\alias{plot_top_species_line}
\title{Plot activity of top species over time using line-style plot}
\usage{
plot_top_species_line(
  dataframe,
  type = c("detections", "sessions", "both"),
  style = c("overlay", "facet"),
  n = 2
)
}
\arguments{
\item{dataframe}{Results dataframe.}

\item{type}{Character. Determines which species to plot: "detections", "sessions", or "both". Default = "detections".
\itemize{
\item "detections": plot species with the highest total number of detections over the deployment.
\item "sessions": plot species detected on the most sessions (regardless of number of detections per night/day).
\item "both": plot species from both criteria. Species may appear in one or both groups. Line type indicates which ranking the species belongs to.
}}

\item{style}{"overlay" or "facet". Default = "overlay".}

\item{n}{Number of species to include (default 2). Maximum 8 species plotted.}
}
\value{
ggplot object.
}
\description{
Plot activity of top species over time using line-style plot
}
\details{
This function internally calls \code{\link{summarise_species_activity}} to determine the total detections and sessions detected per species.
}
\seealso{
\code{\link{summarise_species_activity}}, \code{\link{get_top_species_total}}, \code{\link{get_top_species_sessions}}
}
