% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_classifications_df.R
\name{prepare_classifications_df}
\alias{prepare_classifications_df}
\title{Prepare Classifications Dataframe}
\usage{
prepare_classifications_df(classifications_df)
}
\arguments{
\item{classifications_df}{A dataframe containing classifications. This dataframe
is expected to have an \code{image_path} column from which the deployment ID will be extracted.}
}
\value{
A modified dataframe with two new columns:
\describe{
\item{\code{deployment_id}}{Extracted deployment ID based on the pattern "dep\\d{6}".}
\item{\code{classification_id}}{A unique row number assigned to each classification.}
}
}
\description{
This function processes a classifications dataframe by extracting the deployment ID
from the \code{image_path} column and adding a unique classification ID for each row.
}
