% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_to_standardised_format.R
\name{wrangle_birdnet}
\alias{wrangle_birdnet}
\title{Internal helper: wrangle BirdNET bird results}
\usage{
wrangle_birdnet(results, ...)
}
\description{
Internal helper: wrangle BirdNET bird results
}
\keyword{internal}
