% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_to_standardised_format.R
\name{wrangle_moth}
\alias{wrangle_moth}
\title{Internal helper: wrangle moth results}
\usage{
wrangle_moth(results, ...)
}
\description{
Internal helper: wrangle moth results
}
\keyword{internal}
