% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_to_standardised_format.R
\name{wrangle_results}
\alias{wrangle_results}
\title{Wrangle the data into a standardised format}
\usage{
wrangle_results(
  results_file,
  taxa,
  filepath_colname,
  sitename_colname,
  latitude_colname,
  longitude_colname,
  topspeciespred_colname,
  speciesconf_colname
)
}
\arguments{
\item{results_file}{A string specifying the name of the file containing the concatenated csv results files}

\item{taxa}{A string specifying the name of the taxa being analysed - e.g., moth, bird, bat}

\item{filepath_colname}{A string specifying the name of the column containing filepaths}

\item{sitename_colname}{A string specifying the name of the column containing site names}

\item{latitude_colname}{A string specifying the name of the column containing latitudes}

\item{longitude_colname}{A string specifying the name of the column containing longitudes}

\item{topspeciespred_colname}{A string specifying the name of the column containing top species predictions}

\item{speciesconf_colname}{A string specifying the name of the column containing species confidence scores}
}
\value{
A data.frame containing the wrangled results.
}
\description{
This function takes a concatenated csv file for a deployment and
runs a series of checks and data wrangling steps to format the data in
a standardised way suitable for the downstream analysis pipeline.
Currently just set-up for moth results (AMI/LepiSense)
}
