% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object_store_files.R
\name{download_object_store_files}
\alias{download_object_store_files}
\title{Download files from an S3 bucket}
\usage{
download_object_store_files(
  bucket,
  deployment_id,
  data_type,
  filename,
  download_path = "./downloads",
  credentials_path = "./credentials.json",
  save_download_log = TRUE
)
}
\arguments{
\item{bucket}{Character. Name of the S3 bucket e.g., gbr}

\item{deployment_id}{Character vector or single value. Deployment ID corresponding to each file.
If you are uncertain of this, see the image_path of your classifications, as the ID should be in the path.}

\item{data_type}{Character vector or single value. Data type(s) corresponding to each file.
E.g., snapshot_images, motion_images, ultrasound_recordings, or audible_recordings}

\item{filename}{Character vector. List of basename filenames to download.}

\item{download_path}{Character. Local directory where files will be downloaded. Default is "./downloads".}

\item{credentials_path}{Character. Path to credentials file JSON.}

\item{save_download_log}{Logical. Whether to save the download log.}
}
\value{
A data frame containing deployment IDs, data types, local file paths, and download status. The download status is either, 'successfully downloaded' meaning the file has been downloaded to the download directory, 'already downloaded (skipped)' meaning the file already exists in the download directory, or 'failed download (see warnings)'. This means that there was an error that interrupted the download process. Please check the last warnings with warnings() for more details.
}
\description{
This function downloads files from an S3 bucket based on provided vectors of deployment IDs,
data types, and filenames. The files are downloaded in parallel with the function outputting a log with the final download status of each file.
}
