% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object_store_files.R
\name{download_object_store_files}
\alias{download_object_store_files}
\title{Download files from an S3 bucket}
\usage{
download_object_store_files(
  bucket,
  deployment_id,
  data_type,
  filename,
  classification_id = NULL,
  download_path = "./downloads",
  credentials_path = "./credentials.json",
  save_download_log = TRUE
)
}
\arguments{
\item{bucket}{Character. Name of the S3 bucket.}

\item{deployment_id}{Character vector. List of deployment IDs.}

\item{data_type}{Character vector or single value. Data type(s) corresponding to each file.}

\item{filename}{Character vector. List of filenames to download.}

\item{classification_id}{Numeric or character vector. List of classification IDs.}

\item{download_path}{Character. Local directory where files will be downloaded. Default is "./downloads".}

\item{credentials_path}{Character. Path to credentials file JSON.}

\item{save_download_log}{Logical. Whether to save download log.}
}
\value{
A data frame containing deployment IDs, data types, local file paths, and classification IDs of downloaded files.
}
\description{
This function downloads files from an S3 bucket given a list of deployment IDs,
data types, filenames, and classification IDs. Files are downloaded in parallel.
}
