% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datetimes.R
\name{parse_AMI_datetimes}
\alias{parse_AMI_datetimes}
\title{Extract and parse datetime from filenames}
\usage{
parse_AMI_datetimes(file_names)
}
\arguments{
\item{file_names}{A character vector of filenames containing datetime patterns.}
}
\value{
A vector of parsed datetime objects in UTC timezone.
}
\description{
This function extracts datetime strings from filenames, removes any underscores,
and converts them into datetime objects.
}
\details{
\itemize{
\item The function looks for datetime patterns in the format \code{YYYYMMDD_HHMMSS} or \code{YYYYMMDDHHMMSS}.
\item It removes any underscores found in the extracted datetime strings.
\item Uses \code{lubridate::parse_date_time()} to convert strings into datetime objects.
}
}
\examples{

# Example filenames
file_names <- c("dep000092/snapshot_images/20240918032339-snapshot.jpg", "dep000050/ultrasound_recordings/19700101_000114.WAV")

# Parse datetimes
parsed_dates <- parse_AMI_datetimes(file_names)

}
