% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_RC_rules.R
\name{remove_implausible_predictions}
\alias{remove_implausible_predictions}
\title{Remove implausaible rows based on recording Confidence Rules}
\usage{
remove_implausible_predictions(dataframe_with_rc)
}
\arguments{
\item{classifications_df}{A dataframe containing classifications with record confidence rules appended. This dataset assumes the data is in the format obtained from the classification pipeline on JASMIN.}
}
\value{
The updated classifications dataframe with rows with implausible predictions removed.
}
\description{
This function removes rows in the dataframe where the prediction is implausible (based on location and/or date)
}
